add player movement and physics based on blocks around the player

+++ app/src/main/kotlin/game/camera/Camera.kt
@@ -4,8 +4,8 @@ import org.joml.Vector3d
 
 class Camera(
     val position: Vector3d,
-    var pitch: Double,
-    var yaw: Double,
-    var roll: Double
+    var pitch: Float,
+    var yaw: Float,
+    var roll: Float
 ) {
 }
